extract geometry
