--- rlib-common/src/main/java/javasabr/rlib/common/geom/DirectionType.java
@@ -1,33 +1,0 @@
-package javasabr.rlib.common.geom;
-
-/**
- * The list od direction types.
- *
- * @author JavaSaBr
- */
-public enum DirectionType {
-  /**
-   * Left direction type.
-   */
-  LEFT,
-  /**
-   * Up direction type.
-   */
-  UP,
-  /**
-   * Front direction type.
-   */
-  FRONT,
-  /**
-   * Right direction type.
-   */
-  RIGHT,
-  /**
-   * Down direction type.
-   */
-  DOWN,
-  /**
-   * Behind direction type.
-   */
-  BEHIND
-}
