--- rlib-common/src/main/java/javasabr/rlib/common/geom/bounding/BoundingType.java
@@ -1,21 +1,0 @@
-package javasabr.rlib.common.geom.bounding;
-
-/**
- * The list of bounding types.
- *
- * @author JavaSaBr
- */
-public enum BoundingType {
-  /**
-   * Axis aligned box bounding type.
-   */
-  AXIS_ALIGNED_BOX,
-  /**
-   * Sphere bounding type.
-   */
-  SPHERE,
-  /**
-   * Empty bounding type.
-   */
-  EMPTY,
-}
