+++ rlib-geometry/build.gradle
@@ -1,0 +1,4 @@
+
+dependencies {
+  api projects.rlibCommon
+}
