+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/DirectionType.java
@@ -1,0 +1,13 @@
+package javasabr.rlib.geometry;
+
+/**
+ * @author JavaSaBr
+ */
+public enum DirectionType {
+  LEFT,
+  UP,
+  FRONT,
+  RIGHT,
+  DOWN,
+  BEHIND
+}
