+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/Matrix3f.java
@@ -1,26 +1,32 @@
-package javasabr.rlib.common.geom;
+package javasabr.rlib.geometry;
 
 import javasabr.rlib.common.util.pools.Reusable;
-import org.jspecify.annotations.NullMarked;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.experimental.Accessors;
+import lombok.experimental.FieldDefaults;
 
 /**
- * The implementation of float matrix 3x3.
- *
  * @author JavaSaBr
  */
-@NullMarked
+@Getter
+@Accessors(fluent = true)
+@FieldDefaults(level = AccessLevel.PROTECTED)
 public final class Matrix3f implements Reusable {
 
-  public static final Matrix3f ZERO = new Matrix3f(0, 0, 0, 0, 0, 0, 0, 0, 0);
+  public static final Matrix3f ZERO = new Matrix3f(
+      0, 0, 0,
+      0, 0, 0,
+      0, 0, 0);
 
   public static final Matrix3f IDENTITY = new Matrix3f();
 
   /**
    * Values.
    */
-  protected float val_0_0, val_0_1, val_0_2;
-  protected float val_1_0, val_1_1, val_1_2;
-  protected float val_2_0, val_2_1, val_2_2;
+  float val_0_0, val_0_1, val_0_2;
+  float val_1_0, val_1_1, val_1_2;
+  float val_2_0, val_2_1, val_2_2;
 
   public Matrix3f() {
     val_0_1 = val_0_2 = val_1_0 = val_1_2 = val_2_0 = val_2_1 = 0;
@@ -61,9 +67,6 @@ public final class Matrix3f implements Reusable {
     this.val_2_2 = val_2_2;
   }
 
-  /**
-   * Change all values to absolute.
-   */
   public void absoluteLocal() {
     val_0_0 = Math.abs(val_0_0);
     val_0_1 = Math.abs(val_0_1);
@@ -76,13 +79,6 @@ public final class Matrix3f implements Reusable {
     val_2_2 = Math.abs(val_2_2);
   }
 
-  /**
-   * Multiply the vector by this matrix.
-   *
-   * @param vector the source vector.
-   * @param result the result vector.
-   * @return the result.
-   */
   public Vector3f mult(Vector3f vector, Vector3f result) {
 
     float x = vector.x;
@@ -96,19 +92,6 @@ public final class Matrix3f implements Reusable {
     return result;
   }
 
-  /**
-   * Set the matrix values.
-   *
-   * @param val_0_0 the val 0 0
-   * @param val_0_1 the val 0 1
-   * @param val_0_2 the val 0 2
-   * @param val_1_0 the val 1 0
-   * @param val_1_1 the val 1 1
-   * @param val_1_2 the val 1 2
-   * @param val_2_0 the val 2 0
-   * @param val_2_1 the val 2 1
-   * @param val_2_2 the val 2 2
-   */
   public void set(
       float val_0_0,
       float val_0_1,
@@ -130,19 +113,14 @@ public final class Matrix3f implements Reusable {
     this.val_2_2 = val_2_2;
   }
 
-  /**
-   * Set values from the rotation.
-   *
-   * @param rotation the rotation.
-   * @return this updated matrix.
-   */
   public Matrix3f set(Quaternion4f rotation) {
     return rotation.toRotationMatrix(this);
   }
 
   @Override
   public String toString() {
-    return val_0_0 + ", " + val_0_1 + ", " + val_0_2 + "\n" + val_1_0 + ", " + val_1_1 + ", " + val_1_2 + "\n" + val_2_0
-        + ", " + val_2_1 + ", " + val_2_2 + "\n";
+    return val_0_0 + ", " + val_0_1 + ", " + val_0_2 + "\n" +
+        val_1_0 + ", " + val_1_1 + ", " + val_1_2 + "\n" +
+        val_2_0 + ", " + val_2_1 + ", " + val_2_2 + "\n";
   }
 }
