+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/bounding/BoundingType.java
@@ -1,0 +1,10 @@
+package javasabr.rlib.geometry.bounding;
+
+/**
+ * @author JavaSaBr
+ */
+public enum BoundingType {
+  AXIS_ALIGNED_BOX,
+  SPHERE,
+  EMPTY
+}
