+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/bounding/impl/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.geometry.bounding.impl;
+
+import org.jspecify.annotations.NullMarked;
