+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/bounding/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.geometry.bounding;
+
+import org.jspecify.annotations.NullMarked;
