+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.geometry;
+
+import org.jspecify.annotations.NullMarked;
