+++ rlib-geometry/src/main/java/javasabr/rlib/geometry/util/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.rlib.geometry.util;
+
+import org.jspecify.annotations.NullMarked;
