+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/impl/BasePluginSystem.java
@@ -26,7 +26,6 @@ import javasabr.rlib.common.util.FileUtils;
 import javasabr.rlib.common.util.Utils;
 import javasabr.rlib.common.util.array.Array;
 import javasabr.rlib.common.util.array.ReadOnlyArray;
-import javasabr.rlib.common.util.array.impl.ReadOnlyFastArray;
 import javasabr.rlib.common.util.dictionary.ObjectDictionary;
 import javasabr.rlib.logger.api.Logger;
 import javasabr.rlib.logger.api.LoggerManager;
