+++ settings.gradle
@@ -11,3 +11,4 @@ include ':rlib-logger-api'
 include ':rlib-logger-impl'
 include ':rlib-logger-slf4j'
 include ':rlib-plugin-system'
+include ':rlib-geometry'
